splits up visualizers components
