+++ libs/visualizer/src/components/subject.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { BoxWithText } from './box-with-text';
+
+export function Subject(props) {
+  return (
+    <BoxWithText
+      text={`${props.name}: ${props.value}`}
+      {...props}
+      boxColor="red"
+      onClick={() => props.onClick('subject', props.name)}
+    />
+  );
+}
